requirements.txt
fastapi
sqlalchemy
python-dotenv
validate_email
pydantic
python_jose
passlib
